backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
boto3==1.34.131
botocore==1.34.131
anthropic==0.29.0
docker==7.1.0
websockets==12.0
pydantic==2.7.4
pydantic-settings==2.3.4
httpx==0.27.0
python-dotenv==1.0.1
uuid==1.30
pytz==2024.1
